converted to a library format, updated docker to point to a tag, generally simplified
