--- frameworks/Rust/faf/src/const_config.rs
@@ -1,5 +1,0 @@
-pub const MAX_EPOLL_EVENTS_RETURNED: usize = 4096;
-pub const MAX_EPOLL_CONN: usize = 32674;
-pub const MAX_HEADERS_TO_PARSE: usize = 16;
-pub const REQ_RES_BUFF_SIZE: usize = 4096;
-pub const EPOLL_TIMEOUT_MILLIS: isize = 3000;
